move write and read contacts functions to helpers folder
