+++ models/contacts/getContactById.js
@@ -1,4 +1,4 @@
-import { readContacts } from "./readContacts.js";
+import { readContacts } from "../../helpers/index.js";
 
 export const getContactById = async (contactId) => {
   const contacts = await readContacts();
