+++ models/contacts/index.js
@@ -1,6 +1,3 @@
-export * from "./readContacts.js";
-export * from "./writeContacts.js";
-
 export * from "./listContacts.js";
 export * from "./getContactById.js";
 export * from "./addContact.js";
