+++ models/contacts/updateContact.js
@@ -1,4 +1,5 @@
-import { readContacts, writeContacts, getContactById } from "./index.js";
+import { getContactById } from "./index.js";
+import { readContacts, writeContacts } from "../../helpers/index.js";
 
 export const updateContact = async (contactId, body) => {
   const contacts = await readContacts();
